+++ packages/xchain-bitcoin/__mocks__/haskoin.ts
@@ -1,7 +1,4 @@
-import axios from 'axios'
-import MockAdapter from 'axios-mock-adapter'
-
-const mock = new MockAdapter(axios)
+import mock from './axios-adapter'
 
 type MockConfig = {
   url?: string
@@ -10,11 +7,13 @@ type MockConfig = {
 export default {
   restore: mock.restore,
   init: () => {
-    //Mock https://api.haskoin.com/haskoin-store/btc/address/{address}/balance
+    //Mock GET https://{haskoinurl}/{btc|btctest}/address/{address}/balance
     mock.onGet(/\/address\/\w+\/balance/).reply((config: MockConfig) => {
       const address = config.url?.split('/')?.[6] ?? ''
       const resp = require(`./response/balances/haskoin-${address}.json`)
       return [200, resp]
     })
+    //Mock POST https://{haskoinurl}/{btc|btctest}/transactions
+    mock.onPost(/\/transactions/).reply(() => [200, { txid: 'mock-txid' }])
   },
 }